requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.49.0
imageio==2.34.0
kiwisolver==1.4.5
lazy_loader==0.3
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python-headless==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
PeakUtils==1.3.4
pillow==10.2.0
pyarrow==15.0.0
pycairo==1.26.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2024.1
scikit-image==0.22.0
scipy==1.12.0
setuptools==69.1.0
shapely==2.0.3
six==1.16.0
tifffile==2024.2.12
tzdata==2024.1
wheel==0.42.0
